+++ server/src/sys/terrain.rs
@@ -10,7 +10,7 @@ use common::{
     event::{EventBus, ServerEvent},
     generation::get_npc_name,
     npc::NPC_NAMES,
-    terrain::TerrainGrid,
+    terrain::{SerializedTerrainChunk, TerrainGrid},
     LoadoutBuilder, SkillSetBuilder,
 };
 use common_ecs::{Job, Origin, Phase, System};
@@ -92,6 +92,28 @@ impl<'a> System<'a> for Sys {
 
             // Add to list of chunks to send to nearby players.
             new_chunks.push((key, Arc::clone(&chunk)));
+
+            // Send the chunk to all nearby players.
+            let mut lazy_msg = None;
+            for (presence, pos, client) in (&presences, &positions, &clients).join() {
+                let chunk_pos = terrain.pos_key(pos.0.map(|e| e as i32));
+                // Subtract 2 from the offset before computing squared magnitude
+                // 1 since chunks need neighbors to be meshed
+                // 1 to act as a buffer if the player moves in that direction
+                let adjusted_dist_sqr = (chunk_pos - key)
+                    .map(|e: i32| (e.abs() as u32).saturating_sub(2))
+                    .magnitude_squared();
+
+                if adjusted_dist_sqr <= presence.view_distance.pow(2) {
+                    if lazy_msg.is_none() {
+                        lazy_msg = Some(client.prepare(ServerGeneral::TerrainChunkUpdate {
+                            key,
+                            chunk: Ok(SerializedTerrainChunk::from_chunk(&*chunk)),
+                        }));
+                    }
+                    lazy_msg.as_ref().map(|ref msg| client.send_prepared(&msg));
+                }
+            }
 
             // TODO: code duplication for chunk insertion between here and state.rs
             // Insert the chunk into terrain changes